requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.23
pyodbc==4.0.39
pydantic>=2.6.1
python-dotenv>=1.0.0
pytest>=8.0.0
PyMuPDF==1.23.7
uvicorn==0.24.0
pyyaml>=6.0.1
cryptography>=42.0.0

cachetools>=5.3.0

Brotli>=1.0.9

prometheus-client>=0.16.0

sqlalchemy>=2.0.0
asyncpg>=0.27.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0